Extract http.Request related code to its own source file (http.go)
